+++ src/main/java/de/maxhenkel/pipez/utils/DummyFluidHandler.java
@@ -1,0 +1,49 @@
+package de.maxhenkel.pipez.utils;
+
+import net.minecraftforge.fluids.FluidStack;
+import net.minecraftforge.fluids.capability.IFluidHandler;
+
+import javax.annotation.Nonnull;
+
+public class DummyFluidHandler implements IFluidHandler {
+
+    public static final DummyFluidHandler INSTANCE = new DummyFluidHandler();
+
+    @Override
+    public int getTanks() {
+        return 0;
+    }
+
+    @Nonnull
+    @Override
+    public FluidStack getFluidInTank(int tank) {
+        return FluidStack.EMPTY;
+    }
+
+    @Override
+    public int getTankCapacity(int tank) {
+        return 0;
+    }
+
+    @Override
+    public boolean isFluidValid(int tank, @Nonnull FluidStack stack) {
+        return false;
+    }
+
+    @Override
+    public int fill(FluidStack resource, FluidAction action) {
+        return 0;
+    }
+
+    @Nonnull
+    @Override
+    public FluidStack drain(FluidStack resource, FluidAction action) {
+        return FluidStack.EMPTY;
+    }
+
+    @Nonnull
+    @Override
+    public FluidStack drain(int maxDrain, FluidAction action) {
+        return FluidStack.EMPTY;
+    }
+}